+++ src/tensors/tensor.rs
@@ -18,7 +18,7 @@ pub trait Tensor<T: CoordinateSystem> {
 	
 	/// Returns the coordinate of the tensor, at indices specified by the slice.
 	/// The length of the slice (the number of indices) has to be compatible with the rank of the tensor. 
-	fn get_coord(&self, i: &[u8]) -> T::CoordType;
+	fn get_coord(&self, i: &[usize]) -> T::CoordType;
 	
 	/// Returns the rank of the tensor, that is, the list of the index types.
 	/// A vector would return vec![Contravariant], a metric tensor: vec![Covariant, Covariant].
